shop/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.1
django-bootstrap-form==3.4
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
flake8==5.0.4
flake8-builtins==1.5.3
flake8-django==1.1.5
flake8-import-order==0.18.1
flake8-print==5.0.0
idna==3.4
kombu==5.2.4
mccabe==0.7.0
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
